demo/requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.0.1
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mpmath==1.1.0
networkx==2.2
numpy==1.16.2
Pillow==6.0.0
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
sympy==1.3
urllib3==1.24.1
Werkzeug==0.15.2
